Fixed quite a few compiler warnings. Implemented the constructor for X86VirtualAddressSpace and X64VirtualAddressSpace. A new VirtualAddressSpace should be created via VirtualAddressSpace::create().

+++ src/system/kernel/core/processor/x64/StackFrame.cc
@@ -60,7 +60,7 @@ void X64StackFrame::construct(ProcessorState &state,
   va_list list;
   va_start(list, nParams);
   
-  for(int i = 0; i < nParams; i++)
+  for(size_t i = 0; i < nParams; i++)
   {
     uintptr_t arg = va_arg(list, uintptr_t);
     NOTICE("Setting " << Hex << i << ", " << arg);
